+++ CHANGELOG.md
@@ -7,7 +7,8 @@
 - Drop obsolete `GetSpecifiedInstance` method from `IReadOnlyNamedOptions` and `IWritableNamedOptions` (use `GetInstance` instead)
 - Drop obsolete `FormatProviderBase.LoadConfiguration<T>(WritableOptionsConfiguration<T>)` overloads (use `LoadConfiguration(Type, IWritableOptionsConfiguration)` instead)
 - Drop obsolete `FormatProviderBase.SaveAsync<T>(T, WritableOptionsConfiguration<T>, CancellationToken)` overload (use `SaveAsync<T>(T, IWritableOptionsConfiguration, CancellationToken)` instead)
-- Drop `OnChangeThrottleMs` property (use `OnChangeThrottle` with `TimeSpan` instead)
+- Drop `OnChangeThrottleMs` property (use `OnChangeDebounce` with `TimeSpan` instead)
+- Rename `OnChangeThrottle` to `OnChangeDebounce`
 - Rename `IFormatProvider` to `IWritableFormatProvider` to avoid conflict with `System.IFormatProvider`
 
 ## [0.5.0] - 2026-07-11